Implemented proper locking for zipping/unzipping.
Zips to be unzipped are opened with non-blocking shared locks, and zips
to be created are opened with non-blocking exclusive locks. Failure to
grab the given lock is considered a failure, and errno will likely
contain EWOULDBLOCK. Locking of the multiple components of a multi-
volume zip file is not properly supported.