Solved truncation-before-lock problem with zip locking
Solved problem where concurrent zip writing could cause problems even
when locking is used. The issue was that opening a file for writing
would append it before locking was performed. The solution is to use
mode 'r+b', allowing read/write, with truncation, but you can only open 'r+b'
if the file already exists. Therefore, I added a loop with an fopen
with mode 'ab' (to create the file), then fopen with 'r+b' (to actually
working with the file). ftruncate() isn't called until a lock is gotten,
and even if another process/thread deletes the file after we create it, we'll
just go around in the loop and do it again.