make inttypes.h optional
`inttypes.h` isn't available in certain environments, but instead of
relying on CMake to create a fake local copy, let's handle the matter
in the source code, and let it work with any build system. If
`HAVE_INTTYPES_H` is defined, `inttypes.h` will be used, otherwise
the required values will be defined locally.