Revert "Fixed improper null termination in mz_os_get_file_date. #601"
This reverts commit fe36527157ff96226f5309a99c806475ee0b75d2, which
changed the POSIX mz_os_get_file_date to always remove the last
character from the filename before calling stat() (regardless of whether
the filename ended in a slash).

NUL termination was correct before that change, because "len+1" always
includes the NUL byte at the end of the input string (unless you somehow
have a string that consumes the entire address space and len+1
overflows, but then where would the rest of the program data reside?).