Create parent directory before extracting symlink
At least when using Info-ZIP 3.0, symlinks to both files and directories
lack a trailing slash, so the second mz_path_remove_filename is
unnecessary (and could break things if the parent directory doesn't
precede the symlink in the file). If other implementations do add a
trailing slash, the if statement should handle that.